Rename setupTests.tsx to testUtils.tsx
