+++ src/pages/notFoundPage/notFoundPage.test.tsx
@@ -1,5 +1,5 @@
 import { describe, test, expect } from 'vitest'
-import { renderWithRouter } from '../../support/setupTests'
+import { renderWithRouter } from '../../support/testUtils'
 import paths from '../../routing/paths'
 import NotFoundPage from './notFoundPage'
