+++ src/support/testUtils.tsx
@@ -1,9 +1,16 @@
 import { type ReactElement } from 'react'
 import { BrowserRouter } from 'react-router-dom'
 import { render } from '@testing-library/react'
+import createFetchMock from 'vitest-fetch-mock'
 import { LoginContext } from '../contexts/loginContext'
 import { User } from 'firebase/auth'
 import testProfileImg from './testProfileImg.png'
+
+/*
+ *
+ * Global Test Data
+ *
+ */
 
 export const requireLogin = () => { /* noop */ }
 
@@ -15,6 +22,12 @@ export const testUser = {
   emailVerified: true,
   getIdToken: () => new Promise<string>((resolve, _reject) => resolve('xxxxxxx'))
 } as User
+
+/*
+ *
+ * Test Renderers
+ *
+ */
 
 export const renderWithRouter = (ui: ReactElement) =>
   render(<BrowserRouter>{ui}</BrowserRouter>)
